Feature/contextual metadata discounts (#16)
Following up on our conversations, we have moved the Personalize Discounts into using the "contextual metadata" feature to identify whether a product was being offered at a discount when it was viewed. Therefore, we use a single campaign and get different recommendations depending on the context of whether a discount is active. This will replace the previously mentioned solution. This change involves:

-Changes to training data (user-product interactions) to add the context field which will record which offer is active in the interaction (from {'DISCOUNT':'No'} and {'DISCOUNT':'Yes'})
-Remove the discounts campaign.
-Update the reranking campaign training and inference in the Lambda function that automatically trains all Personalize campaigns to take the contextual metadata (we use the reranking campaign with discounts because it allows us to “pick the top N”) (reranking also supports contextual metadata).
-Update the recommendations discounts endpoint to provide the context to Personalize.
-Update the experimentations framework to match this.
Ran unit tests.
-Workshop: We will also update the Jupyter notebook we added to show:
--How the interactions are generated.
--How the campaign is created.
--How recommendations might be retrieved from Personalize and from the endpoint.
--We also visualise to ensure that the recommender gives expected results.
By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

This work was done by Dae.mn Team.
Co-authored-by: Akram Dweikat akram@akramdweikat.com
Co-authored-by: Akram Dweikat akram.dweikat@dae.mn
Co-authored-by: Damien Jade Duff damien.duff@daemonsolutions.com
Co-authored-by: Joe Major joe.major@dae.mn