Update Amazon Pay sign-in + UI stub
To support the frictionless Alexa checkout Amazon Pay session signing has been integrated into the Retail Demo Store.
To showcase this, we allow users to enter their Amazon Pay details in CloudFormation and get an Amazon Pay button on the Retail Demo Store.
The button is a stub and if the Amazon Pay sandbox details were entered correctly, after starting the session will redirect to the Amazon Pay documentation.
Upcoming work to be delivered the Alexa skill will allow full checkout in the Alexa sandbox.