feat: add new flags to stage.sh
When working locally, we don't want to make s3 upload public or repackage all code everytime. Currently, we have to comment some lines out during development and may accidentally commit and push it. This change adds a script to accept new flags to change the behaviour to suit local development.