Fix duplicate pinpoint events (#450)
The Analytics class registers itself with Amplify, which then does a call back to the configure method.   The Analytics configure method adds the AWSPinpointProvider if there  are no pluggables, which there won't be on initialization.  Therefore, the AWSPinpointProvider will always be added by default, and so adding it again, as it has been done in main.js, causes duplicate event processing.